preparing for paper release (#77)
* updating the tmp/readme

* updating precommit

* fixing cmd line interface in __main__

* updating typing on sigmoid fit

* adding logging re how CV is performed

* initial topn modeling implementation. This will store the results in a separate output directory to differentiate from all data modeling

* initial attempt of stage 3 of modeling. This seems to run without issues based on limited testing.

* stepwise modelling

* this separates the sigmoid step 3 into its own function; reoganizes cmd line input into reusable groups; sets the evaluate_interactor_significance estimator to LinearRegression by default

* removing windows 2019 from CI

* Update tfbpmodeling/loop_modeling.py


* Update tfbpmodeling/loop_modeling.py


* Update tfbpmodeling/loop_modeling.py


* with pre-commit

* removing islice from bootstrap loop

* removing fixtures to conftest; fixing spacing in loop module; adding pytest to CI

* only running tests on ubuntu

* trying to configure codecov

* debugging codecov

* debugging codecov

* removing codecov badge and updating the pytest badge

* adding the codecov badge back in -- secret corrected in the repo

* debugging codecov

* still debugging codecov

* Add cubic ptf and standardization (#29)

* init

* After installing pre-commit

* add center and scaling

* changing the names for center scaling

* Update tfbpmodeling/__main__.py


* Update tfbpmodeling/tests/test_lasso_modeling.py


* Update tfbpmodeling/tests/test_lasso_modeling.py


* Update tfbpmodeling/tests/test_lasso_modeling.py


* Update tfbpmodeling/lasso_modeling.py


* Update tfbpmodeling/lasso_modeling.py


* debugging the names for scale and center

* editing annotatioins

---------


* Set random state on bootstraps; Remove unweighted bootstrap option (#31)

* intermediate

* propogating bootstrappedmodelinginputdata changes to __main__

* removed unweighted bootstrap options

* removing top_n as argparse option from step3 sigmoid parser

* adding center_scale to argparse

* setting drop_intercept to True permanently for sigmoid worker

* the sigmoid parameters must have args.drop_intercept still

* handling intercepts

* fixing typo in center_scale logging

* changing the way the formula is logged

* removing truncation from formula logging

* adding logging on random_state in bootstrappedmodelinput

* removing sample weight cv log

* removing sample weight cv logging

* Add function to exclude all model variables (#33)

* adding function to exclude all predictors. this can be used to exclude all and then use args.add_model_variables to customize the formula

* casting predictor_variables to list

* fixing centering and scaling (#36)

* loop exits if no variable selected within the loop, fixes issue 34 (#37)

* loop exits if no variable selected within the loop, fixes issue 34

* fixing linter issues

* linter issues

---------


* setting evaluate_interactor_significance ci_level to the parse args argument

* Tommy new stage3 (#41)

* WIP

* fixing imported but not used

* adding argument to include stage4_lasso

* changing formatting problems

* adding logger infomation for stage 4 method

* changing location of logger info for stage 4 method

* modifying logger.info to evaluate_interactor_significance

* adding f string to Writing the final interactor significance results to {output_significance_file}

* fixing logging

---------


* fixing error in stratification_classification that reversed the bin_by_binding_only param (#44)

* adding feature stage4_topn (#43)

* adding feature stage4_topn

* modifying argument parser

* Align response_df with predictors from get_modeling_data to ensure consistency when top_n_masked is enabled

* Align response_df with predictors from get_modeling_data to ensure consistency when top_n_masked is enabled

* aligning stratified_cv_r2

* restoring changes to evaluate_interactor_significance_linear

* commit after fixing error in stratification_classification

* attempt to fixing inconsistent numbers of samples

* fixing pr

---------


* Add row max in interactor significance (#48)

* setting row max depending on model variables

* setting row max depending on model variables

* adding testing on log for evaluate_interactor_significance

* removing ptf from all data formula by default (#51)

* Remove bin by binding and Add check on number of features and increase minimal test case size (#54)

* saving changes for remove_bin_by_binding

* Add check on number of features and increase minimal test case size

* Refactor main by adding interface.py (#56)

* create interface.py to separate main

* add tests to verify important logging statements

* move the logging configuration back into main

* create interface.py to separate main

* add tests to verify important logging statements

* move the logging configuration back into main

* fixing interface

* removing bin_by_binding_only from test arguments to interface

* adding name == main to main script

* create interface.py to separate main

* move the logging configuration back into main

* removing bin_by_binding_only from test arguments to interface

* rebasing refactor onto dev

* adding calling to main

* adding a feature column in test_interface

---------


* Separate the functions and objects in lasso_modeling.py (#58)

* create interface.py to separate main

* add tests to verify important logging statements

* move the logging configuration back into main

* create interface.py to separate main

* add tests to verify important logging statements

* move the logging configuration back into main

* fixing interface

* removing bin_by_binding_only from test arguments to interface

* adding name == main to main script

* create interface.py to separate main

* move the logging configuration back into main

* removing bin_by_binding_only from test arguments to interface

* rebasing refactor onto dev

* adding calling to main

* adding a feature column in test_interface

* separate the functions and objects in lasso_modeling.py

* Refactor main by adding interface.py (#56)

* create interface.py to separate main

* add tests to verify important logging statements

* move the logging configuration back into main

* create interface.py to separate main

* add tests to verify important logging statements

* move the logging configuration back into main

* fixing interface

* removing bin_by_binding_only from test arguments to interface

* adding name == main to main script

* create interface.py to separate main

* move the logging configuration back into main

* removing bin_by_binding_only from test arguments to interface

* rebasing refactor onto dev

* adding calling to main

* adding a feature column in test_interface

---------


* create interface.py to separate main

move the logging configuration back into main

fixing interface

create interface.py to separate main

move the logging configuration back into main

removing bin_by_binding_only from test arguments to interface

rebasing refactor onto dev

create interface.py to separate main

move the logging configuration back into main

removing bin_by_binding_only from test arguments to interface

adding name == main to main script

separate the functions and objects in lasso_modeling.py

* renaming loop_modeling

* separate the tests out into files

* separate the tests out into files

* fixing evaluate_interactor_significance_lassocv

---------


* removing scale_center from interface (#60)

---------
